feat: 1.拆分api接口和轮询检查服务功能;2.增加优雅关闭和资源清理
