+++ src/core/mod.rs
@@ -1,8 +1,10 @@
+pub mod api;
 pub mod collector;
 pub mod doctor;
 pub mod ent;
 pub mod logger;
-pub use collector::{listen, ServiceChecker};
+pub use api::*;
+pub use collector::ServiceChecker;
 pub use doctor::*;
 pub use ent::*;
 pub use logger::*;
